Refactor: 계산식 훅으로 분리
